requirements.txt
django==1.9.4
django-taggit==0.18.0
